requirements.txt
asgiref==3.7.2
boto3==1.28.53
botocore==1.31.53
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
Django==4.2.5
djangorestframework==3.14.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
moto==4.2.4
packaging==23.1
pluggy==1.3.0
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
responses==0.23.3
s3transfer==0.6.2
serverless-wsgi==3.0.2
six==1.16.0
sqlparse==0.4.4
types-PyYAML==6.0.12.12
urllib3==1.26.16
Werkzeug==2.3.7
xmltodict==0.13.0
